add license and documentation for imgaug module
