--- modules/aug/include/opencv2/aug.hpp
@@ -1,9 +1,0 @@
-#ifndef OPENCV_AUG_HPP
-#define OPENCV_AUG_HPP
-
-#include "opencv2/aug/transforms.hpp"
-#include "opencv2/aug/transforms_det.hpp"
-#include "opencv2/aug/functional.hpp"
-#include "opencv2/aug/rng.hpp"
-
-#endif
