--- modules/aug/src/precomp.hpp
@@ -1,10 +1,0 @@
-#ifndef OPENCV_AUG_PRECOMP_H
-#define OPENCV_AUG_PRECOMP_H
-
-#include "opencv2/aug.hpp"
-#include <cstdlib>
-#include <ctime>
-#include <cmath>
-#include <algorithm>
-
-#endif
