--- modules/aug/test/test_main.cpp
@@ -1,3 +1,0 @@
-#include "test_precomp.hpp"
-
-CV_TEST_MAIN("cv")
