--- modules/aug/test/test_precomp.hpp
@@ -1,10 +1,0 @@
-#ifndef __OPENCV_TEST_PRECOMP_HPP__
-#define __OPENCV_TEST_PRECOMP_HPP__
-
-#include "opencv2/ts.hpp"
-#include "opencv2/aug.hpp"
-
-static uint64 seed=0;
-static cv::RNG rng(seed);
-
-#endif
