+++ modules/imgaug/CMakeLists.txt
@@ -1,0 +1,2 @@
+set(the_description "Data Augmentation Module")
+ocv_define_module(imgaug opencv_imgproc opencv_core opencv_imgcodecs opencv_highgui WRAP python)
