+++ modules/imgaug/include/opencv2/imgaug.hpp
@@ -1,0 +1,19 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+#ifndef OPENCV_IMGAUG_HPP
+#define OPENCV_IMGAUG_HPP
+
+#include "opencv2/imgaug/transforms.hpp"
+#include "opencv2/imgaug/transforms_det.hpp"
+#include "opencv2/imgaug/functional.hpp"
+#include "opencv2/imgaug/rng.hpp"
+
+/** @defgroup imgaug Data Augmentation Module for Efficient Data Preprocessing
+ *  @{
+ *  @defgroup det Data Augmentation for Object Detection
+ *  @}
+*/
+
+
+#endif
