+++ modules/imgaug/src/precomp.hpp
@@ -1,0 +1,13 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+#ifndef OPENCV_AUG_PRECOMP_H
+#define OPENCV_AUG_PRECOMP_H
+
+#include "opencv2/imgaug.hpp"
+#include <cstdlib>
+#include <ctime>
+#include <cmath>
+#include <algorithm>
+
+#endif
