+++ modules/imgaug/test/test_precomp.hpp
@@ -1,0 +1,13 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+#ifndef __OPENCV_TEST_PRECOMP_HPP__
+#define __OPENCV_TEST_PRECOMP_HPP__
+
+#include "opencv2/ts.hpp"
+#include "opencv2/imgaug.hpp"
+
+static uint64 seed=0;
+static cv::RNG rng(seed);
+
+#endif
